@@ -76,16 +81,19 @@
     "rimraf": "^2.6.2",
     "selenium-webdriver": "^4.0.0-alpha.1",
     "style-loader": "^0.21.0",
-    "webpack": "^4.12.0",
+    "typescript": "^3.4.1",
+    "webpack": "^4.29.6",
     "webpack-bundle-analyzer": "^3.0.3",
     "webpack-cli": "^3.0.7",
     "webpack-dev-server": "^3.1.4"
   },
   "dependencies": {
+    "@babel/polyfill": "^7.4.3",
+    "antd": "^3.17.0",
     "babel-polyfill": "^6.26.0",
-    "babel-watch": "^2.0.7",
-    "chart.js": "^2.7.2",
+    "chart.js": "^2.8.0",
     "concurrently": "^4.0.1",
+    "event-source-polyfill": "^1.0.5",
     "fullcalendar-reactwrapper": "^1.0.7",
     "jquery": "^3.3.1",
     "leaflet": "^1.4.0",
