@@ -1,8 +1,8 @@
-import React from "react"
-import _Select2 from "react-select2-wrapper"
-import $ from "jquery"
-import Loader from "aurigauikit/components/Loader"
-import "./style.css"
+import React from 'react'
+import _Select2 from 'react-select2-wrapper'
+// import $ from 'jquery'
+import Loader from 'aurigauikit/components/Loader'
+import './style.css'
 
 class Select2 extends React.Component {
   constructor(props) {
@@ -12,7 +12,7 @@ class Select2 extends React.Component {
   static defaultProps = {
     data: [],
     willDisplay: element => element,
-    didSelect: element => element
+    didSelect: element => element,
   }
 
   contains(selected) {
@@ -36,38 +36,40 @@ class Select2 extends React.Component {
 
   onUnselect(e) {
     const options = Array.from(e.target.selectedOptions)
-    if (this.props.multiple)
+    if (this.props.multiple) {
       this.props.didSelect(
         this.props.value.filter(v => options.find(o => o.value === this.props.willDisplay(v)))
       )
+    }
   }
 
   render() {
     const {
-      didSelect, //eslint-disable-line no-unused-vars
+      didSelect, // eslint-disable-line no-unused-vars
       willDisplay,
       data,
       value,
       loading,
+      createOption, // eslint-disable-line no-unused-vars
       ...props
     } = this.props
     return [
       <_Select2
         key="select"
         {...props}
-        onClick={() => setTimeout(() => $(".select2-search__field").focus(), 50)}
-        data={data.map(v => (v ? willDisplay(v) : ""))}
+        // onClick={() => setTimeout(() => $(".select2-search__field").focus(), 50)}
+        data={data.map(v => (v ? willDisplay(v) : ''))}
         value={
           value
-            ? typeof value !== "string" && value.length > 0
+            ? typeof value !== 'string' && value.length > 0
               ? value.map(v => willDisplay(v))
               : willDisplay(value)
-            : ""
+            : ''
         }
         onSelect={e => this.onSelect(e)}
         onUnselect={(e, a) => this.onUnselect(e, a)}
       />,
-      loading && <Loader key="loader" style={{ marginTop: -28, marginRight: 8 }} />
+      loading && <Loader key="loader" style={{ marginTop: -28, marginRight: 8 }} />,
     ]
   }
 }
