@@ -1,2 +1,1 @@
-import Select2 from "./Select2ReWrapper"
-export default Select2
+export { default } from './SelectAntDesign'
