@@ -1,8 +1,11 @@
-import "./css/ext/select2.css"
-import "./css/form-select.css"
-import "./css/ext/normalize.min.css"
-import "./css/simplegrid.css"
-import "./css/ext/font-awesome.min.css"
-import "./css/ext/tooltipster.bundle.min.css"
-import "./css/OpenSans.css"
-import "./main.less"
+import './css/ext/select2.css'
+import './css/form-select.css'
+import './css/ext/normalize.min.css'
+import './css/simplegrid.css'
+// import "./css/ext/font-awesome.min.css"
+import './fontawesome/css/all.min.css'
+import './fontawesome/css/v4-shims.min.css'
+import './css/ext/tooltipster.bundle.min.css'
+import './css/OpenSans.css'
+import './main.less'
+import './ant-override.css'
