@@ -1,185 +1,200 @@
-import { lazy } from "react"
-import ChartMetadata from "../components/Chart/meta.json"
-import CheckboxMetadata from "../components/Checkbox/meta.json"
-import OrganizationalUnitMetadata from "../components/OrganizationalUnit/meta.json"
-import RadioMetadata from "../components/Radio/meta.json"
-import SimpleTableMetadata from "../components/SimpleTable/meta.json"
+/** AUTO-GENERATED - do not modify */
+import { lazy } from 'react'
+import ChartMetadata from '../components/Chart/meta.json'
+import CheckboxMetadata from '../components/Checkbox/meta.json'
+import OrganizationalUnitMetadata from '../components/OrganizationalUnit/meta.json'
+import RadioMetadata from '../components/Radio/meta.json'
+import SimpleTableMetadata from '../components/SimpleTable/meta.json'
 export default {
   Chart: {
     metadata: ChartMetadata,
     Bar: {
-      component: lazy(() => import("../components/Chart/Bar")),
-      code: () => import("!raw-loader!../components/Chart/Bar")
+      component: lazy(() => import('../components/Chart/Bar')),
+      code: () => import('!raw-loader!../components/Chart/Bar'),
     },
     Doughnut: {
-      component: lazy(() => import("../components/Chart/Doughnut")),
-      code: () => import("!raw-loader!../components/Chart/Doughnut")
+      component: lazy(() => import('../components/Chart/Doughnut')),
+      code: () => import('!raw-loader!../components/Chart/Doughnut'),
     },
     Line: {
-      component: lazy(() => import("../components/Chart/Line")),
-      code: () => import("!raw-loader!../components/Chart/Line")
-    }
+      component: lazy(() => import('../components/Chart/Line')),
+      code: () => import('!raw-loader!../components/Chart/Line'),
+    },
   },
   Checkbox: {
     metadata: CheckboxMetadata,
     Base: {
-      component: lazy(() => import("../components/Checkbox/Base")),
-      code: () => import("!raw-loader!../components/Checkbox/Base")
-    }
+      component: lazy(() => import('../components/Checkbox/Base')),
+      code: () => import('!raw-loader!../components/Checkbox/Base'),
+    },
   },
   ContextMenu: {
     metadata: {},
     Base: {
-      component: lazy(() => import("../components/ContextMenu/Base")),
-      code: () => import("!raw-loader!../components/ContextMenu/Base")
+      component: lazy(() => import('../components/ContextMenu/Base')),
+      code: () => import('!raw-loader!../components/ContextMenu/Base'),
     },
     Hidden: {
-      component: lazy(() => import("../components/ContextMenu/Hidden")),
-      code: () => import("!raw-loader!../components/ContextMenu/Hidden")
-    }
+      component: lazy(() => import('../components/ContextMenu/Hidden')),
+      code: () => import('!raw-loader!../components/ContextMenu/Hidden'),
+    },
   },
   DatePicker: {
     metadata: {},
     Base: {
-      component: lazy(() => import("../components/DatePicker/Base")),
-      code: () => import("!raw-loader!../components/DatePicker/Base")
+      component: lazy(() => import('../components/DatePicker/Base')),
+      code: () => import('!raw-loader!../components/DatePicker/Base'),
     },
     Clearable: {
-      component: lazy(() => import("../components/DatePicker/Clearable")),
-      code: () => import("!raw-loader!../components/DatePicker/Clearable")
-    }
+      component: lazy(() => import('../components/DatePicker/Clearable')),
+      code: () => import('!raw-loader!../components/DatePicker/Clearable'),
+    },
   },
   Loader: {
     metadata: {},
     Base: {
-      component: lazy(() => import("../components/Loader/Base")),
-      code: () => import("!raw-loader!../components/Loader/Base")
-    }
+      component: lazy(() => import('../components/Loader/Base')),
+      code: () => import('!raw-loader!../components/Loader/Base'),
+    },
   },
   Maps: {
     metadata: {},
     Simple: {
-      component: lazy(() => import("../components/Maps/Simple")),
-      code: () => import("!raw-loader!../components/Maps/Simple")
-    }
+      component: lazy(() => import('../components/Maps/Simple')),
+      code: () => import('!raw-loader!../components/Maps/Simple'),
+    },
   },
   OrganizationalUnit: {
     metadata: OrganizationalUnitMetadata,
     Base: {
-      component: lazy(() => import("../components/OrganizationalUnit/Base")),
-      code: () => import("!raw-loader!../components/OrganizationalUnit/Base")
+      component: lazy(() => import('../components/OrganizationalUnit/Base')),
+      code: () => import('!raw-loader!../components/OrganizationalUnit/Base'),
     },
     Default: {
-      component: lazy(() => import("../components/OrganizationalUnit/Default")),
-      code: () => import("!raw-loader!../components/OrganizationalUnit/Default")
+      component: lazy(() => import('../components/OrganizationalUnit/Default')),
+      code: () => import('!raw-loader!../components/OrganizationalUnit/Default'),
     },
     Filter: {
-      component: lazy(() => import("../components/OrganizationalUnit/Filter")),
-      code: () => import("!raw-loader!../components/OrganizationalUnit/Filter")
-    }
+      component: lazy(() => import('../components/OrganizationalUnit/Filter')),
+      code: () => import('!raw-loader!../components/OrganizationalUnit/Filter'),
+    },
   },
   Radio: {
     metadata: RadioMetadata,
     Base: {
-      component: lazy(() => import("../components/Radio/Base")),
-      code: () => import("!raw-loader!../components/Radio/Base")
-    }
+      component: lazy(() => import('../components/Radio/Base')),
+      code: () => import('!raw-loader!../components/Radio/Base'),
+    },
+  },
+  ReactTable: {
+    metadata: {},
+    Basic: {
+      component: lazy(() => import('../components/ReactTable/Basic')),
+      code: () => import('!raw-loader!../components/ReactTable/Basic'),
+    },
   },
   Select2: {
     metadata: {},
     Base: {
-      component: lazy(() => import("../components/Select2/Base")),
-      code: () => import("!raw-loader!../components/Select2/Base")
+      component: lazy(() => import('../components/Select2/Base')),
+      code: () => import('!raw-loader!../components/Select2/Base'),
     },
     Complex: {
-      component: lazy(() => import("../components/Select2/Complex")),
-      code: () => import("!raw-loader!../components/Select2/Complex")
-    }
+      component: lazy(() => import('../components/Select2/Complex')),
+      code: () => import('!raw-loader!../components/Select2/Complex'),
+    },
   },
   SimpleTable: {
     metadata: SimpleTableMetadata,
     Base: {
-      component: lazy(() => import("../components/SimpleTable/Base")),
-      code: () => import("!raw-loader!../components/SimpleTable/Base")
+      component: lazy(() => import('../components/SimpleTable/Base')),
+      code: () => import('!raw-loader!../components/SimpleTable/Base'),
     },
     Controlled: {
-      component: lazy(() => import("../components/SimpleTable/Controlled")),
-      code: () => import("!raw-loader!../components/SimpleTable/Controlled")
+      component: lazy(() => import('../components/SimpleTable/Controlled')),
+      code: () => import('!raw-loader!../components/SimpleTable/Controlled'),
     },
     DynamicContents: {
-      component: lazy(() => import("../components/SimpleTable/DynamicContents")),
-      code: () => import("!raw-loader!../components/SimpleTable/DynamicContents")
+      component: lazy(() => import('../components/SimpleTable/DynamicContents')),
+      code: () => import('!raw-loader!../components/SimpleTable/DynamicContents'),
     },
     Empty: {
-      component: lazy(() => import("../components/SimpleTable/Empty")),
-      code: () => import("!raw-loader!../components/SimpleTable/Empty")
+      component: lazy(() => import('../components/SimpleTable/Empty')),
+      code: () => import('!raw-loader!../components/SimpleTable/Empty'),
     },
     Pageable: {
-      component: lazy(() => import("../components/SimpleTable/Pageable")),
-      code: () => import("!raw-loader!../components/SimpleTable/Pageable")
+      component: lazy(() => import('../components/SimpleTable/Pageable')),
+      code: () => import('!raw-loader!../components/SimpleTable/Pageable'),
     },
     Sortable: {
-      component: lazy(() => import("../components/SimpleTable/Sortable")),
-      code: () => import("!raw-loader!../components/SimpleTable/Sortable")
-    }
+      component: lazy(() => import('../components/SimpleTable/Sortable')),
+      code: () => import('!raw-loader!../components/SimpleTable/Sortable'),
+    },
   },
   Slider: {
     metadata: {},
     Advanced: {
-      component: lazy(() => import("../components/Slider/Advanced")),
-      code: () => import("!raw-loader!../components/Slider/Advanced")
+      component: lazy(() => import('../components/Slider/Advanced')),
+      code: () => import('!raw-loader!../components/Slider/Advanced'),
     },
     Base: {
-      component: lazy(() => import("../components/Slider/Base")),
-      code: () => import("!raw-loader!../components/Slider/Base")
-    }
+      component: lazy(() => import('../components/Slider/Base')),
+      code: () => import('!raw-loader!../components/Slider/Base'),
+    },
   },
   Spinner: {
     metadata: {},
     Base: {
-      component: lazy(() => import("../components/Spinner/Base")),
-      code: () => import("!raw-loader!../components/Spinner/Base")
-    }
+      component: lazy(() => import('../components/Spinner/Base')),
+      code: () => import('!raw-loader!../components/Spinner/Base'),
+    },
+  },
+  StaticPanel: {
+    metadata: {},
+    Base: {
+      component: lazy(() => import('../components/StaticPanel/Base')),
+      code: () => import('!raw-loader!../components/StaticPanel/Base'),
+    },
   },
   Switch: {
     metadata: {},
     Base: {
-      component: lazy(() => import("../components/Switch/Base")),
-      code: () => import("!raw-loader!../components/Switch/Base")
-    }
+      component: lazy(() => import('../components/Switch/Base')),
+      code: () => import('!raw-loader!../components/Switch/Base'),
+    },
   },
   Ticker: {
     metadata: {},
     Base: {
-      component: lazy(() => import("../components/Ticker/Base")),
-      code: () => import("!raw-loader!../components/Ticker/Base")
+      component: lazy(() => import('../components/Ticker/Base')),
+      code: () => import('!raw-loader!../components/Ticker/Base'),
     },
     Url: {
-      component: lazy(() => import("../components/Ticker/Url")),
-      code: () => import("!raw-loader!../components/Ticker/Url")
-    }
+      component: lazy(() => import('../components/Ticker/Url')),
+      code: () => import('!raw-loader!../components/Ticker/Url'),
+    },
   },
   Workflow: {
     metadata: {},
     Base: {
-      component: lazy(() => import("../components/Workflow/Base")),
-      code: () => import("!raw-loader!../components/Workflow/Base")
+      component: lazy(() => import('../components/Workflow/Base')),
+      code: () => import('!raw-loader!../components/Workflow/Base'),
     },
     Complete: {
-      component: lazy(() => import("../components/Workflow/Complete")),
-      code: () => import("!raw-loader!../components/Workflow/Complete")
+      component: lazy(() => import('../components/Workflow/Complete')),
+      code: () => import('!raw-loader!../components/Workflow/Complete'),
     },
     CustomDeleted: {
-      component: lazy(() => import("../components/Workflow/CustomDeleted")),
-      code: () => import("!raw-loader!../components/Workflow/CustomDeleted")
+      component: lazy(() => import('../components/Workflow/CustomDeleted')),
+      code: () => import('!raw-loader!../components/Workflow/CustomDeleted'),
     },
     Deleted: {
-      component: lazy(() => import("../components/Workflow/Deleted")),
-      code: () => import("!raw-loader!../components/Workflow/Deleted")
+      component: lazy(() => import('../components/Workflow/Deleted')),
+      code: () => import('!raw-loader!../components/Workflow/Deleted'),
     },
     Empty: {
-      component: lazy(() => import("../components/Workflow/Empty")),
-      code: () => import("!raw-loader!../components/Workflow/Empty")
-    }
-  }
+      component: lazy(() => import('../components/Workflow/Empty')),
+      code: () => import('!raw-loader!../components/Workflow/Empty'),
+    },
+  },
 }
